// src/types/ebay.ts
export interface EbayConfig {
  clientId: string;
  clientSecret: string;
  redirectUri?: string;
  locale?: LocaleEnum;
  marketplaceId?: string;
  contentLanguage?: string;
  environment: 'production' | 'sandbox';
  accessToken?: string;
  refreshToken?: string;
  appAccessToken?: string;
}
// ...
export interface EbayApiError {
  errors: {
    errorId: number;
    domain: string;
    category: string;
    message: string;
    longMessage?: string;
    parameters?: {
      name: string;
      value: string;
    }[];
  }[];
}